Revert "fixed server encryption"
This reverts commit 3f197a1ae9508f8d9099b1ce76be4b968a027262.

+++ include/crypt.h
@@ -1,8 +1,8 @@
 #ifndef CRYPT
 #define CRYPT
 
-void encrypt(unsigned int k[], unsigned int tmp[]);
+void encrypt(unsigned long k[], unsigned long tmp[]);
 void encrypt_Handler(SerializedMessage_t * sermes);
-void decrypt(unsigned int k[], unsigned int tmp[]);
+void decrypt(unsigned long k[], unsigned long tmp[]);
 void decrypt_Handler(char msg[], int size);
 #endif // CRYPT
